+++ src/app/models/speed-test-settings.model.ts
@@ -1,0 +1,6 @@
+import {SpeedTestFileModel} from './speed-test-file.model';
+
+export class SpeedTestSettingsModel {
+  public iterations?: number;
+  public file?: SpeedTestFileModel;
+}