+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0.0]
 ### Changed
+* Model for handling custom settings by [jrquick17](https://github.com/jrquick17)
+
+### Changed
 * Prefixed models with speed-test by [jrquick17](https://github.com/jrquick17)
 
 ## [Released]